Refactor code organization and enhance abstraction
- LazyGroceryGrid: Updated to accept a list of grocery groups. Each group now includes an optional title and a list of groceries. This modification removes dependencies on specific group identifiers, such as booleans or enums.

- GroceryList: Made public for broader usage throughout the app, ensuring a consistent UI experience.

- Directory Renaming: Changed "feature" directory to "screens" for clarity. This adjustment aims to align package naming with the typical structure where packages within this module generally represent a single screen. "Screens" provides a clearer and more consistent naming convention.

- Reorganize files into directories for better organization.
