+++ app/src/main/java/com/rendox/grocerygenius/MainActivity.kt
@@ -10,7 +10,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.rendox.grocerygenius.feature.grocery_list.GroceryListRoute
+import com.rendox.grocerygenius.screens.grocery_list.screen.GroceryListRoute
 import com.rendox.grocerygenius.ui.theme.GroceryGeniusTheme
 
 class MainActivity : ComponentActivity() {
